project-1/requirements.txt
Flask>=2.2.2
mysql-connector-python==8.0.28
